+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and 
 ### Fixed
 
 - Board import failing with "Method POST not allowed" due to manually set Content-Type header stripping the multipart boundary
+- Board Members dialog can now be closed by pressing Escape (closes #70)
 
 ### Added